moving class to models directory more optimization
