+++ Gateway/crds-angular/Controllers/API/ProfileController.cs
@@ -3,6 +3,7 @@ using System.Web;
 using System.Web.Http;
 using System.Web.Http.Description;
 using crds_angular.Models;
+using crds_angular.Models.Crossroads.Serve;
 using crds_angular.Security;
 using crds_angular.Services;
