+++ Gateway/crds-angular/Models/Crossroads/Serve/ServeRole.cs
@@ -1,0 +1,10 @@
+using Newtonsoft.Json;
+
+namespace crds_angular.Models.Crossroads.Serve
+{
+    public class ServeRole
+    {
+        [JsonProperty(PropertyName = "name")]
+        public string Name { get; set; }
+    }
+}
