+++ Gateway/crds-angular/crds-angular.csproj
@@ -223,6 +223,13 @@
     <Compile Include="Models\Crossroads\FamilyMember.cs" />
     <Compile Include="Models\Crossroads\ServeOccurance.cs" />
     <Compile Include="Models\Crossroads\GroupOpportunity.cs" />
+    <Compile Include="Models\Crossroads\Serve\EventType.cs" />
+    <Compile Include="Models\Crossroads\Serve\ServeEvent.cs" />
+    <Compile Include="Models\Crossroads\Serve\ServeRole.cs" />
+    <Compile Include="Models\Crossroads\Serve\ServingDay.cs" />
+    <Compile Include="Models\Crossroads\Serve\ServingTeam.cs" />
+    <Compile Include="Models\Crossroads\Serve\ServingTime.cs" />
+    <Compile Include="Models\Crossroads\Serve\TeamMember.cs" />
     <Compile Include="Models\Crossroads\ServingRSVP.cs" />
     <Compile Include="Models\Crossroads\ServingTeam.cs" />
     <Compile Include="Models\Crossroads\Skill.cs" />
